dashboard 최적화
