+++ src/views/PlaceholderView.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="dashboard-placeholder">
+    <p>대시보드 페이지입니다. (준비 중)</p>
+  </div>
+</template>
+
+<style scoped>
+.dashboard-placeholder {
+  flex: 1;
+  padding: 40px 32px;
+  font-size: 16px;
+  color: var(--color-text-default);
+}
+</style>
